refactor(ui): split `reports` into more files
- split out pure function into its own file
  - to make clear that it is pure and unrelated to React state, context, etc

- split out `ReportsFilters` component similar to `CronWorkflowFilters`, `WorkflowFilters`, etc
  - be consistent with how all these work
  - similar to 25e1939e25551cd15d89bd47e4232c8073b40a9c, refactor logic to have an `onChange` function and retrieve Workflows in a top-level effect (after all state has been set)
    - and similarly add an SCSS file with the appropriate styles
      - note: this may be worth refactoring as nearly all the filters have the same styles

- refactor `ReportsFilters` logic to use async/await syntax
  - also refactor `CronWorkflowList` to use newer async/await syntax consistent with future state of codebase
