+++ ui/src/app/reports/components/reports-filters.scss
